feat: Implement webhook event schema using Zod
This commit introduces a new event schema using Zod for validating webhook events from Facebook and TikTok.

The changes include:

- Created `src/modules/events/common.ts` to define common schemas and utility functions for event handling.
- Created `src/modules/events/facebook.ts` to define the schema for Facebook events.
- Created `src/modules/events/tiktok.ts` to define the schema for TikTok events.
- Modified `src/modules/events/event.zod.ts` to import and combine the Facebook and TikTok event schemas into a discriminated union for a unified event schema.
